renamed field to be more generic
